% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticities.R
\name{elasticities_single_prophet_model}
\alias{elasticities_single_prophet_model}
\title{Currently only works with a single external regressor and model factory}
\usage{
elasticities_single_prophet_model(workflow, elasticities_df,
  group_vars = c("PRODUCT", "SHOP"), mod_name = "1__1",
  joining_vars = c("PRODUCT"), elasticity_var = "PRICE_ELASTICITY")
}
\arguments{
\item{workflow}{model workflow object}

\item{group_vars}{names of model factory grouping variables}

\item{mod_name}{character id of the model in the workflow object}

\item{joining_vars}{named vector of joining variables with the elasticities_df}

\item{elasticity_var}{character name of the elasticity variable in elasticities_df}

\item{elasticties_df}{data.frame containing elasticities data}

\item{evaluation_funs}{list of names of functions to calculate elasticities for. Typically functions from MLmetrics}
}
\value{
data.frame
}
\description{
Currently only works with a single external regressor and model factory
}
