% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_raw_data}
\alias{interpolate_raw_data}
\title{Applies interpolation to the raw data slot in a workflow}
\usage{
interpolate_raw_data(workflow, interp, ...)
}
\arguments{
\item{workflow}{a workflow object}

\item{interp}{named vector with the names being the columns to be interpolated the values are the type of interpolation (see `interpolate_ts_data`)}

\item{\dots}{arguments to be passed to the interpolation function}
}
\value{
a workflow object
}
\description{
Applies interpolation to the raw data slot in a workflow
}
\seealso{
interpolate_ts_data
}
