% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_ts_data}
\alias{interpolate_ts_data}
\title{Interpolate missing days from a daily time series}
\usage{
interpolate_ts_data(ts_dat, interp, ...)
}
\arguments{
\item{ts_dat}{data.frame of ts data with dates to be interpolated}

\item{interp}{named vector with the names being the columns to be interpolated the values are the type of interpolation (see details)}

\item{\dots}{arguments to be passed to the interpolation function}
}
\value{
a data frame with missing values interpolated
}
\description{
This function extends the time series to the full extent of the
minimum and maximum values and interpolated the values in a flexible way.
The data frame must have a ds column of dates
}
