% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition_ts_data}
\alias{partition_ts_data}
\title{Partitions time series data into a test, training and validation sets}
\usage{
partition_ts_data(workflow, test_from, train_p, test_p)
}
\arguments{
\item{workflow}{object of class ModelWorkflow}

\item{train_p}{proportion of the data to go in the training set}

\item{test_p}{proportion of the data to go in the test set}
}
\value{
an object of class ModelWorkflow with an updated data slot
}
\description{
Partitions time series data into a test, training and validation sets
}
