% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_prophet_model_factory}
\alias{validate_prophet_model_factory}
\title{Currently only works with a single external regressor and model factory}
\usage{
validate_prophet_model_factory(workflow, elasticities_df,
  group_vars = c("PRODUCT", "SHOP"), joining_vars = c("PRODUCT"),
  elasticity_var = "PRICE_ELASTICITY", evaluation_funs = list("MAPE",
  "Poisson_LogLoss"), cores)
}
\arguments{
\item{workflow}{model workflow object}

\item{group_vars}{names of model factory grouping variables}

\item{joining_vars}{named vector of joining variables with the elasticities_df}

\item{elasticity_var}{character name of the elasticity variable in elasticities_df}

\item{evaluation_funs}{list of names of functions to calculate elasticities for. Typically functions from MLmetrics}

\item{cores}{number of processor cores to use}

\item{elasticties_df}{data.frame containing elasticities data}
}
\value{
data.frame
}
\description{
Currently only works with a single external regressor and model factory
}
